Add files via upload
We consider a parametric quasi-variational inequality (QVI) without any convexity assumption. Using the concept of optimal value function, we transform the problem into that of solving a nonsmooth system of inequalities. Based on this reformulation, new coderivative estimates as well as robust stability conditions for the optimal solution  map of this QVI are developed. Also, for an optimization problem with QVI constraint, necessary optimality conditions are constructed and subsequently a tailored  semismooth Newton-type method is designed, implemented, and tested on a wide range of optimization examples from the literature. In addition to the fact that our approach does not require convexity, its coderivative and stability analysis does not involve second order derivatives and subsequently, the proposed Newton scheme does not need third order derivatives.